시발 존나많이 고쳤어

+++ baba-snack-homshop/src/main/java/com/babasnack/demo/member/dao/MemberDao.java
@@ -14,7 +14,7 @@ import com.babasnack.demo.entity.Member;
 public interface MemberDao {
     @Insert("INSERT INTO member (username, password, pno_tell, ps_email, join_day) " +
             "VALUES (#{username}, #{password},#{pnoTell}, #{email}, #{joinDay})")
-   public Long save(Member member);
+   public Integer save(Member member);
 
     @Select("SELECT*FROM member")
     public List<Member> findAllForMember();
